NativeExec function for calling raw intrinsic functions on registers

+++ include/FastSIMD/ToolSet/Generic/Scalar/f32x1.h
@@ -125,6 +125,6 @@ namespace FS
     template<FastSIMD::FeatureSet SIMD, typename = EnableIfNative<f32<1, SIMD>>>
     FS_FORCEINLINE f32<1, SIMD> InvSqrt( const f32<1, SIMD>& a )
     {
-        return 1 / std::sqrtf( a.native.f );
+        return 1 / std::sqrt( a.native.f );
     }
 }
